// frontend/bindings/androidfs/internal/model/models.ts
export class FileEntry {
    "name": string;
    "path": string;
    "is_dir": boolean;
    "size": number;
    "mod_time": time$0.Time;

    /**
     * permission bits, e.g. "drwxr-xr-x"
     */
    "mode": string;

    /**
     * symlink target, empty if not a link
     */
    "link": string;

    /** Creates a new FileEntry instance. */
    constructor($$source: Partial<FileEntry> = {}) {
        if (!("name" in $$source)) {
            this["name"] = "";
        }
        if (!("path" in $$source)) {
            this["path"] = "";
        }
        if (!("is_dir" in $$source)) {
            this["is_dir"] = false;
        }
        if (!("size" in $$source)) {
            this["size"] = 0;
        }
        if (!("mod_time" in $$source)) {
            this["mod_time"] = null;
        }
        if (!("mode" in $$source)) {
            this["mode"] = "";
        }
        if (!("link" in $$source)) {
            this["link"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new FileEntry instance from a string or object.
     */
    static createFrom($$source: any = {}): FileEntry {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new FileEntry($$parsedSource as Partial<FileEntry>);
    }
}